moved isAuthenticated to its own file, added authentication to project routes
