+++ api/routes/index.js
@@ -1,19 +1,6 @@
 var express = require('express');
 var router = express.Router();
-
-var isAuthenticated = function (req, res, next) {
-	// if user is authenticated in the session, call the next() to call the next request handler 
-	// Passport adds this method to request object. A middleware is allowed to add properties to
-	// request and response objects
-	if (req.isAuthenticated()){
-		console.log('You are authorized to go to home page');
-		return next();
-	}
-	    
-	// if the user is not authenticated then redirect him to the login page
-	console.log('You are not authorized');
-	res.redirect('/');
-}
+var isAuthenticated = require('../utils/authentication');
 
 module.exports = function(passport){
